python-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sea-bridge"
version = "0.1.0"
description = "SEA Bridge - Document Translation Service"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SEA Bridge Team", email = "contact@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Core
    "fastapi>=0.109.2,<1.0.0",
    "uvicorn[standard]>=0.27.1,<0.28.0",
    "python-multipart>=0.0.9,<0.1.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "pydantic-settings>=2.2.1,<3.0.0",
    "pydantic>=2.6.4,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    
    # AWS
    "boto3>=1.34.19,<2.0.0",
    "aioboto3",
    "botocore>=1.34.19,<2.0.0",
    
    # Document Processing
    "PyMuPDF>=1.23.21,<2.0.0",
    "python-docx>=1.1.0,<2.0.0",
    "Pillow>=10.2.0,<11.0.0",
    "python-magic>=0.4.27,<0.5.0",
    
    # Caching
    "redis>=5.0.1,<6.0.0",
    "hiredis>=2.3.1,<3.0.0",
    
    # Utilities
    "python-slugify>=8.0.1,<9.0.0",
    "python-magic-bin>=0.4.14,<0.5.0; sys_platform == 'win32'",
    "python-magic>=0.4.27,<0.5.0; sys_platform != 'win32'",
    
    # Async
    "anyio>=4.3.0,<5.0.0",
    "httpx>=0.27.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.2,<9.0.0",
    "pytest-asyncio>=0.23.5,<0.24.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-mock>=3.12.0,<4.0.0",
    
    # Code Quality
    "black>=24.1.1,<25.0.0",
    "isort>=5.13.2,<6.0.0",
    "mypy>=1.8.0,<2.0.0",
    "pylint>=3.0.3,<4.0.0",
    
    # Documentation
    "docstring-to-markdown>=0.12,<0.13",
    "sphinx>=7.2.6,<8.0.0",
    "sphinx-rtd-theme>=2.0.0,<3.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "auto"